+++ sql/core/src/main/scala/org/apache/spark/sql/DataFrameWriter.scala
@@ -747,7 +747,7 @@ final class DataFrameWriter[T] private[sql](ds: Dataset[T]) {
     val bucketing =
       getBucketSpec.map(spec => CatalogV2Implicits.BucketSpecHelper(spec).asTransform).toSeq
     val clustering = clusteringColumns.map { colNames =>
-      ClusterByTransform(colNames.map(col => FieldReference(col)))
+      ClusterByTransform(colNames.map(FieldReference(_)))
     }
     partitioning ++ bucketing ++ clustering
   }
